backend/requirements.txt
fastapi
uvicorn
pydantic
requests
fastapi
uvicorn
pydantic
requests
httpx
beautifulsoup4
newspaper3k
PyPDF2
pdfplumber
python-docx
tiktoken
chromadb
sqlalchemy
python-multipart
python-dotenv
rich
openai
pydub
vosk